requirements.txt
Django==2.2.4
django-bootstrap3==11.1.0
djangorestframework==3.10.2
pkg-resources==0.0.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0

wheel==0.33.4
